split out files for iso tiles and penrose triangulation
